8207370: Several Monocle robot tests fail because of deprecated assertEquals calls

+++ tests/system/src/test/java/test/robot/com/sun/glass/ui/monocle/TestApplication.java
@@ -365,8 +365,8 @@ public class TestApplication extends Application {
                 TestLogShim.log("y = " + robot.getMouseY());
                 TestLogShim.log("targetX = " + targetX);
                 TestLogShim.log("targetY = " + targetY);
-                Assert.assertEquals(targetX, robot.getMouseX());
-                Assert.assertEquals(targetY, robot.getMouseY());
+                Assert.assertEquals(targetX, (int) robot.getMouseX());
+                Assert.assertEquals(targetY, (int) robot.getMouseY());
             });
             frameWait(1);
         } finally {
